fix gnu size_t bug

+++ uv/bytebuf.cc
@@ -55,9 +55,9 @@ buf_pool_t default_buf_pool;
 
 x_mem_pool default_x_mem_pool;
 
-void *x_malloc(size_t size) {
+void *x_malloc(long unsigned int size) {
     void *res;
-    std::multimap<size_t, void*>::iterator it;
+    std::multimap<long unsigned int, void*>::iterator it;
     it = default_x_mem_pool.ptr_pool.lower_bound(size);
     if(it != default_x_mem_pool.ptr_pool.end()) {
         // INFO("malloc %lu, give %lu, pool_size: %lu", size, (*it).first, default_x_mem_pool.ptr_pool.size());
@@ -72,7 +72,7 @@ void *x_malloc(size_t size) {
 }
 
 void x_free(void *ptr) {
-    size_t size = default_x_mem_pool.ptr_size_map[ptr];
+    long unsigned int size = default_x_mem_pool.ptr_size_map[ptr];
     default_x_mem_pool.ptr_pool.insert(std::make_pair(size, ptr));
     return;
 }
